requirements.txt
django-dotenv==1.4.2
django-filter==2.1.0
django-leaflet==0.24.0
djangorestframework==3.9.4
djangorestframework-gis==0.14
gunicorn==19.9.0
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
